requirements.txt
apache-airflow-providers-postgres
psycopg2-binary
pandas
sqlalchemy
requests